risolto problema subplot immaguini filtrate

+++ ConvFilter/Filter.m
@@ -38,6 +38,6 @@ images{7}=sharp_img;
 figure;
 for i = 1:7
     subplot(2, 4, i);
-    imshow(images{i},[]);
+    imshow(uint8(images{i}),[]);
     title([titles(i)]); 
 end
